혼란을 주는 클래스명 변경 qr_scan_screen.dart -> register_screen.dart   QRScanScreen -> RegisterScreen     포함되어 있던 클래스 QRScanner 별도의 파일로 분리 (new) qr_scan_screen.dart   QRScanner -> QRScanScreen
